Fix/correct orientation on detected objects (#5)
This PR supports usdot-fhwa-stol/carma-platform#1557 by creating a passthrough for the bounding box orientation reported by the Euclidean cluster nodes inside the tracking nodes. Since the orientation is not part of the internal kalman filter it was previously dropped. This PR serves as a workaround to pass it through unfiltered.

* Fix orientation transform for multi_object_tracker to match autoware_auto commit 670f0fae65c34280418bb6adf51fd7cca21b0baf

* Add unfiltered orientation to tracked objects

* Fix build errors

* fixes

Co-authored-by: carma developer <carma@dot.gov>